Report queue depth ratio as a metric to use for autoscaling.
Rename JitterStatisticsCollector to StatisticsCollector.
Update README to reflect property changes.
