+++ src/main/resources/application.properties
@@ -14,7 +14,7 @@ com.tarterware.redis.port=6379
 com.tarterware.roadrunner.vehicle-polling-period=100ms
 com.tarterware.roadrunner.vehicle-update-period=250ms
 
-com.tarterware.roadrunner.jitter-stat-capacity=200
+com.tarterware.roadrunner.queue-stat-capacity=10
 
 spring.profiles.active=dev
